# static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Telefonia API",
    "description": "Telefonia API para gerenciamento de telefones bloqueados",
    "contact": {
      "email": "joaopedroguimaraes96@gmail.com"
    },
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "phones",
      "description": "Telefones bloqueados"
    },
    {
      "name": "verify",
      "description": "Verificador de números bloqueados"
    },
    {
      "name": "api",
      "description": "Requisições sobre a API"
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Status de execução da API",
        "description": "Requisição padrão para verificar o status de execução da API",
        "operationId": "status",
        "responses": {
          "200": {
            "description": "Sucesso, a API está executando normalmente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "string"
                    },
                    "text": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/phones": {
      "get": {
        "tags": [
          "phones"
        ],
        "summary": "Todos os números bloqueados",
        "description": "Retorna todos os números anteriormente inseridos na blocklist",
        "operationId": "getAllPhones",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/blocked_phones"
                  }
                }
              }
            }
          },
          "400": {
            "description": "JSON foi utilizado e de forma inválida",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "phones"
        ],
        "summary": "Bloquear telefones",
        "operationId": "postBlockedPhones",
        "requestBody": {
          "description": "Adicionar um telefone na blocklist",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/blocked_phones"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Telefones bloqueados com sucesso, de acordo com \"blocked_status\" em cada caso retornado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/blocked_phones_result"
                  }
                }
              }
            }
          },
          "400": {
            "description": "JSON inválido ou vazio"
          },
          "500": {
            "description": "Erro inesperado no servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "phones"
        ],
        "summary": "Liberar telefones para o discador",
        "operationId": "deleteBlockedPhones",
        "requestBody": {
          "description": "Remover os telefones listados da blocklist, para que eles não fiquem bloqueados",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/blocked_phones"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Telefones desbloqueados com sucesso, de acordo com \"blocked_status\"",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/unblocked_phones_result"
                  }
                }
              }
            }
          },
          "400": {
            "description": "JSON inválido ou vazio"
          },
          "500": {
            "description": "Erro inesperado no servidor"
          }
        }
      }
    },
    "/verify": {
      "post": {
        "tags": [
          "verify"
        ],
        "summary": "Verificar telefones bloqueados",
        "description": "Verificar os telefones bloqueados de acordo com um JSON que liste todos os números a serem verificados",
        "operationId": "verify",
        "requestBody": {
          "description": "Lista de telefones para verificar o status",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/blocked_phones"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Verificação dos telefones listados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/unblocked_phones_result"
                  }
                }
              }
            }
          },
          "400": {
            "description": "JSON inválido ou vazio"
          },
          "500": {
            "description": "Erro inesperado no servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "blocked_phones": {
        "type": "object",
        "required": [
          "phone_number"
        ],
        "properties": {
          "phone_number": {
            "type": "string",
            "maxLength": 25
          }
        }
      },
      "blocked_phones_result": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string",
            "maxLength": 25
          },
          "blocked_status": {
            "type": "string",
            "enum": [
              "blocked",
              "invalid_phone_number",
              "duplicated"
            ]
          }
        }
      },
      "unblocked_phones_result": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string",
            "maxLength": 25
          },
          "blocked_status": {
            "type": "string",
            "enum": [
              "unblocked",
              "not_found"
            ]
          }
        }
      }
    }
  }
}